# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "fakESB",
    "description": "",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/import": {
      "post": {
        "tags": [
          "crate::http::admin::config"
        ],
        "operationId": "importConfig",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportBundle"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportSummary"
                }
              }
            }
          }
        }
      }
    },
    "/api/logs": {
      "get": {
        "tags": [
          "crate::http::admin::logs"
        ],
        "operationId": "listLogs",
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "method",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "path",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "action",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/RuleAction"
            }
          },
          {
            "name": "status_code",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "start_time",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "end_time",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogPage"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "crate::http::admin::logs"
        ],
        "operationId": "purgeLogs",
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "method",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "path",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "action",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/RuleAction"
            }
          },
          {
            "name": "status_code",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "start_time",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "end_time",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurgeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/logs/{id}": {
      "get": {
        "tags": [
          "crate::http::admin::logs"
        ],
        "operationId": "getLog",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogDetail"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        }
      },
      "delete": {
        "tags": [
          "crate::http::admin::logs"
        ],
        "operationId": "deleteLog",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": ""
          }
        }
      }
    },
...
    "/api/rules": {
      "get": {
        "tags": [
          "crate::http::admin::rules"
        ],
        "operationId": "listRules",
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RulePage"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "crate::http::admin::rules"
        ],
        "operationId": "createRule",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RulePayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/rules/test": {
      "post": {
        "tags": [
          "crate::http::admin::rules"
        ],
        "operationId": "testRule",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleTestPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleTestResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/rules/{id}": {
      "put": {
        "tags": [
          "crate::http::admin::rules"
        ],
        "operationId": "updateRule",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RulePayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdResponse"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        }
      },
      "delete": {
        "tags": [
          "crate::http::admin::rules"
        ],
        "operationId": "deleteRule",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": ""
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BodyFormat": {
        "type": "string",
        "enum": [
          "json",
          "xml",
          "text"
        ]
      },
      "BodyMatcher": {
        "type": "object",
        "required": [
          "format"
        ],
        "properties": {
          "contains": {
            "type": [
              "string",
              "null"
            ]
          },
          "equal_to": {
            "type": [
              "string",
              "null"
            ]
          },
          "fields": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/StringMatcher"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "format": {
            "$ref": "#/components/schemas/BodyFormat"
          },
          "json_equal_to": {},
          "matches": {
            "type": [
              "string",
              "null"
            ]
          },
          "xpath": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/StringMatcher"
            },
            "propertyNames": {
              "type": "string"
            }
          }
        }
      },
      "ExportBundle": {
        "type": "object",
        "properties": {
          "rules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Rule"
            }
          },
          "targets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Target"
            }
          },
          "templates": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResponseTemplate"
            }
          }
        }
      },
      "IdResponse": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ImportSummary": {
        "type": "object",
        "required": [
          "targets_imported",
          "templates_imported",
          "rules_imported"
        ],
        "properties": {
          "rules_imported": {
            "type": "integer",
            "format": "int64"
          },
          "targets_imported": {
            "type": "integer",
            "format": "int64"
          },
          "templates_imported": {
            "type": "integer",
            "format": "int64"
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "LogDetail": {
        "allOf": [
          {
            "$ref": "#/components/schemas/RequestLog"
          },
          {
            "type": "object",
            "required": [
              "snapshots"
            ],
            "properties": {
              "snapshots": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/MessageSnapshot"
                }
              }
            }
          }
        ]
      },
      "LogPage": {
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RequestLog"
            }
          },
          "total": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "MessageSnapshot": {
        "type": "object",
        "required": [
          "id",
          "kind",
          "raw_body",
          "normalized_json"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "kind": {
            "$ref": "#/components/schemas/SnapshotKind"
          },
          "normalized_json": {},
          "raw_body": {
            "type": "string"
          }
        }
      },
      "PurgeResponse": {
        "type": "object",
        "required": [
          "deleted"
        ],
        "properties": {
          "deleted": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "RenderedPreview": {
        "type": "object",
        "required": [
          "status_code",
          "content_type",
          "headers",
          "raw_body"
        ],
        "properties": {
          "content_type": {
            "type": "string"
          },
          "headers": {},
          "raw_body": {
            "type": "string"
          },
          "status_code": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
      "RequestLog": {
        "type": "object",
        "required": [
          "id",
          "occurred_at",
          "method",
          "path",
          "body_format",
          "request_headers",
          "response_headers"
        ],
        "properties": {
          "action": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/RuleAction"
              }
            ]
          },
          "body_format": {
            "$ref": "#/components/schemas/BodyFormat"
          },
          "content_type": {
            "type": [
              "string",
              "null"
            ]
          },
          "error_message": {
            "type": [
              "string",
              "null"
            ]
          },
          "http_status_code": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "latency_ms": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "method": {
            "type": "string"
          },
          "occurred_at": {
            "type": "string",
            "format": "date-time"
          },
          "path": {
            "type": "string"
          },
          "query_string": {
            "type": [
              "string",
              "null"
            ]
          },
          "request_headers": {},
          "response_headers": {},
          "rule_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "target_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          }
        }
      },
      "ResponseTemplate": {
        "type": "object",
        "required": [
          "id",
          "name",
          "content_type",
          "raw_template",
          "format",
          "status_code",
          "headers",
          "enabled",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "content_type": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "enabled": {
            "type": "boolean"
          },
          "format": {
            "type": "string"
          },
          "headers": {},
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "note": {
            "type": [
              "string",
              "null"
            ]
          },
          "raw_template": {
            "type": "string"
          },
          "status_code": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Rule": {
        "type": "object",
        "required": [
          "id",
          "matcher",
          "action",
          "priority",
          "delay_ms",
          "sequence_mode",
          "enabled",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "action": {
            "$ref": "#/components/schemas/RuleAction"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "delay_ms": {
            "type": "integer",
            "format": "int32"
          },
          "enabled": {
            "type": "boolean"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "matcher": {
            "$ref": "#/components/schemas/RuleMatcher"
          },
          "note": {
            "type": [
              "string",
              "null"
            ]
          },
          "priority": {
            "type": "integer",
            "format": "int32"
          },
          "response_template_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "sequence_mode": {
            "type": "boolean"
          },
          "sequence_steps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SequenceStep"
            }
          },
          "target_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "RuleAction": {
        "type": "string",
        "enum": [
          "proxy",
          "static"
        ]
      },
      "RuleMatcher": {
        "type": "object",
        "properties": {
          "body": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/BodyMatcher"
              }
            ]
          },
          "headers": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/StringMatcher"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "method": {
            "type": [
              "string",
              "null"
            ]
          },
          "path": {
            "type": [
              "string",
              "null"
            ]
          },
          "path_pattern": {
            "type": [
              "string",
              "null"
            ]
          },
          "query": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/StringMatcher"
            },
            "propertyNames": {
              "type": "string"
            }
          }
        }
      },
      "RulePage": {
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Rule"
            }
          },
          "total": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "RulePayload": {
        "type": "object",
        "required": [
          "matcher",
          "action"
        ],
        "properties": {
          "action": {
            "$ref": "#/components/schemas/RuleAction"
          },
          "delay_ms": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "enabled": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "matcher": {
            "$ref": "#/components/schemas/RuleMatcher"
          },
          "note": {
            "type": [
              "string",
              "null"
            ]
          },
          "priority": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "response_template_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "sequence_mode": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "sequence_steps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SequenceStepPayload"
            }
          },
          "target_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          }
        }
      },
      "RuleTestPayload": {
        "type": "object",
        "required": [
          "method",
          "path"
        ],
        "properties": {
          "body": {
            "type": [
              "string",
              "null"
            ]
          },
          "content_type": {
            "type": [
              "string",
              "null"
            ]
          },
          "headers": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "method": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "query": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "propertyNames": {
              "type": "string"
            }
          }
        }
      },
      "RuleTestResult": {
        "type": "object",
        "required": [
          "matched"
        ],
        "properties": {
          "action": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/RuleAction"
              }
            ]
          },
          "matched": {
            "type": "boolean"
          },
          "priority": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "rendered": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/RenderedPreview"
              }
            ]
          },
          "rule_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "target_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "target_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "template_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          }
        }
      },
      "SequenceStep": {
        "type": "object",
        "required": [
          "id",
          "rule_id",
          "step_index",
          "template_id"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "rule_id": {
            "type": "integer",
            "format": "int64"
          },
          "step_index": {
            "type": "integer",
            "format": "int32"
          },
          "template_id": {
            "type": "integer",
            "format": "int64"
          },
          "template_name": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "SequenceStepPayload": {
        "type": "object",
        "required": [
          "template_id"
        ],
        "properties": {
          "template_id": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "SnapshotKind": {
        "type": "string",
        "enum": [
          "request",
          "response"
        ]
      },
      "StringMatcher": {
        "type": "object",
        "properties": {
          "contains": {
            "type": [
              "string",
              "null"
            ]
          },
          "equal_to": {
            "type": [
              "string",
              "null"
            ]
          },
          "matches": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "Target": {
        "type": "object",
        "required": [
          "id",
          "name",
          "base_url",
          "enabled",
          "timeout_ms",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "base_url": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "enabled": {
            "type": "boolean"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "note": {
            "type": [
              "string",
              "null"
            ]
          },
          "timeout_ms": {
            "type": "integer",
            "format": "int32"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}